build: fix C++20 compiler warnings

+++ src/driver/gyro/lsm6dso.cpp
@@ -54,7 +54,7 @@ void lsm6dso_configure() {
 uint8_t lsm6dso_read(uint8_t reg) {
   spi_bus_device_reconfigure(&gyro_bus, SPI_MODE_TRAILING_EDGE, SPI_SPEED_SLOW);
 
-  uint8_t buffer[2] = {reg | 0x80, 0x00};
+  uint8_t buffer[2] = {uint8_t(reg | 0x80), 0x00};
 
   const spi_txn_segment_t segs[] = {
       spi_make_seg_buffer(buffer, buffer, 2),
